[00][Edifikana]Enable BE to load config values
