+++ .idea/fileTemplates/Compose Feature.kt.child.1.kt
@@ -1,24 +1,15 @@
 package ${PACKAGE_NAME}
 
-import com.cramsan.framework.core.compose.BaseViewModel
-import com.cramsan.framework.core.compose.ViewModelDependencies
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.MutableSharedFlow
-import kotlinx.coroutines.flow.SharedFlow
-import kotlinx.coroutines.flow.StateFlow
+import androidx.compose.runtime.Composable
+import org.jetbrains.compose.ui.tooling.preview.Preview
 
 /**
- * ViewModel for the ${NAME} screen.
- **/
- // TODO: Register this ViewModel for dependecy injection
-class ${NAME}ViewModel(
-    dependencies: ViewModelDependencies,
-) : BaseViewModel<${NAME}Event, ${NAME}UIState>(
-    dependencies,
-    ${NAME}UIState.Initial,
-    TAG,
-) {
-    companion object {
-        private const val TAG = "${NAME}ViewModel"
-    }
+ * Preview for the ${NAME} feature screen.
+ */
+@Preview
+@Composable
+private fun ${NAME}ScreenPreview() {
+    ${NAME}Content(
+        content = ${NAME}UIState(true),
+    )
 }
