+++ framework/core/src/commonMain/kotlin/com/cramsan/framework/core/UnifiedDispatcherProvider.kt
@@ -1,0 +1,15 @@
+package com.cramsan.framework.core
+
+import kotlinx.coroutines.CoroutineDispatcher
+
+/**
+ * Unified implementation of [DispatcherProvider] that uses the same dispatcher for both IO and UI operations.
+ */
+class UnifiedDispatcherProvider(
+    private val coroutineDispatcher: CoroutineDispatcher,
+) : DispatcherProvider {
+
+    override fun ioDispatcher() = coroutineDispatcher
+
+    override fun uiDispatcher() = coroutineDispatcher
+}
